new file: parseFilesFromMessageContent
