+++ hooks/use-deploy.ts
@@ -3,6 +3,7 @@
 import { toast } from '@/components/toast';
 import { Message } from 'ai';
 import { useState, useEffect } from 'react';
+import { parseFilesFromMessageContent } from '@/lib/vercel/parseFilesFromMessageContent';
 
 export interface DeployContextValue {
   deploy: (message: Message) => Promise<any>;
@@ -22,20 +23,6 @@ export function useDeploy(accessToken?: string | null): DeployContextValue {
   useEffect(() => {
     if (result || error || isLoading) setShowOverlay(true);
   }, [result, error, isLoading]);
-
-  // Helper to extract code blocks with filenames from LLM output
-  function parseFilesFromMessageContent(content: string): Array<{ file: string; data: string; encoding: string }> {
-    const fileBlocks: Array<{ file: string; data: string; encoding: string }> = [];
-    if (!content) return fileBlocks;
-    // Regex to match code blocks like ```tsx file="app/page.tsx" ...code...```
-    const codeBlockRegex = /```[a-zA-Z]+ file=\"([^\"]+)\"\n([\s\S]*?)```/g;
-    let match;
-    while ((match = codeBlockRegex.exec(content)) !== null) {
-      const [, file, data] = match;
-      fileBlocks.push({ file, data: data.trim(), encoding: 'utf-8' });
-    }
-    return fileBlocks;
-  }
 
   async function deploy(message: Message) {
     if (!accessToken) return;
